requirements.txt
PyGithub==2.1.1
PyYAML==6.0.1
requests==2.31.0
colorama==0.4.6 